refactor mint into several components
