+++ cashu/mint/protocols.py
@@ -1,0 +1,8 @@
+from typing import Protocol
+
+from ..core.base import MintKeyset, MintKeysets
+
+
+class SupportsKeysets(Protocol):
+    keyset: MintKeyset
+    keysets: MintKeysets
